# Faster CPU TG for GQA models